refactor to remove dependencies to extended module class (because module class may not be extended if state is broken and should not be extended because it is marked as internal)
